+++ fixtures/ios_app_with_static_libraries/Modules/B/Sources/B.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+public class B {
+
+    public static let value: String = "bValue"
+
+    public static func printFromB() {
+        print("print from B")
+    }
+
+}