+++ fixtures/ios_app_with_static_libraries/Sources/AppDelegate.swift
@@ -1,0 +1,27 @@
+import UIKit
+import A
+
+@UIApplicationMain
+class AppDelegate: UIResponder, UIApplicationDelegate {
+
+    var window: UIWindow?
+
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        window = UIWindow(frame: UIScreen.main.bounds)
+        let viewController = UIViewController()
+        viewController.view.backgroundColor = .white
+        window?.rootViewController = viewController
+        window?.makeKeyAndVisible()
+
+        A.printFromA()
+
+        return true
+    }
+
+}
+
+public class AClassInThisBundle {
+
+  public static let value: String = "aValue"
+
+}